解决了大视频下载报错的问题

+++ BiliDuang/JSONCallback/Player.cs
@@ -15,7 +15,7 @@ namespace BiliDuang.JSONCallback.Player
         /// <summary>
         /// 
         /// </summary>
-        public int size { get; set; }
+        public long size { get; set; }
         /// <summary>
         /// 
         /// </summary>
